requirements.txt
Pillow
pytesseract
requests
beautifulsoup4